+++ src/buffer_cache/buf_patch.hpp
@@ -19,7 +19,7 @@ public:
 
     // Serializes the patch to the given destination address
     void serialize(char* destination) const;
-    size_t get_serialized_size() const;
+    uint16_t get_serialized_size() const;
 
     patch_counter_t get_patch_counter() const;
     ser_transaction_id_t get_transaction_id() const;
@@ -34,7 +34,7 @@ protected:
     // These are for usage in subclasses
     buf_patch_t(const block_id_t block_id, const patch_counter_t patch_counter, const ser_transaction_id_t applies_to_transaction_id, const patch_operation_code_t operation_code);
     virtual void serialize_data(char* destination) const = 0;
-    virtual size_t get_data_size() const = 0;
+    virtual uint16_t get_data_size() const = 0;
 
     static const patch_operation_code_t OPER_MEMCPY = 0;
     static const patch_operation_code_t OPER_MEMMOVE = 1;
@@ -48,8 +48,8 @@ private:
 
 class memcpy_patch_t : public buf_patch_t {
 public:
-    memcpy_patch_t(const block_id_t block_id, const patch_counter_t patch_counter, const ser_transaction_id_t applies_to_transaction_id, const size_t dest_offset, const char *src, const size_t n);
-    memcpy_patch_t(const block_id_t block_id, const patch_counter_t patch_counter, const ser_transaction_id_t applies_to_transaction_id, const char* data, const size_t data_length);
+    memcpy_patch_t(const block_id_t block_id, const patch_counter_t patch_counter, const ser_transaction_id_t applies_to_transaction_id, const uint16_t dest_offset, const char *src, const uint16_t n);
+    memcpy_patch_t(const block_id_t block_id, const patch_counter_t patch_counter, const ser_transaction_id_t applies_to_transaction_id, const char* data, const uint16_t data_length);
 
     virtual ~memcpy_patch_t();
 
@@ -57,29 +57,29 @@ public:
 
 protected:
     virtual void serialize_data(char* destination) const;
-    virtual size_t get_data_size() const;
+    virtual uint16_t get_data_size() const;
 
 private:
     char* src_buf;
-    size_t dest_offset;
-    size_t n;
+    uint16_t dest_offset;
+    uint16_t n;
 };
 
 class memmove_patch_t : public buf_patch_t {
 public:
-    memmove_patch_t(const block_id_t block_id, const patch_counter_t patch_counter, const ser_transaction_id_t applies_to_transaction_id, const size_t dest_offset, const size_t src_offset, const size_t n);
-    memmove_patch_t(const block_id_t block_id, const patch_counter_t patch_counter, const ser_transaction_id_t applies_to_transaction_id, const char* data, const size_t data_length);
+    memmove_patch_t(const block_id_t block_id, const patch_counter_t patch_counter, const ser_transaction_id_t applies_to_transaction_id, const uint16_t dest_offset, const uint16_t src_offset, const uint16_t n);
+    memmove_patch_t(const block_id_t block_id, const patch_counter_t patch_counter, const ser_transaction_id_t applies_to_transaction_id, const char* data, const uint16_t data_length);
 
     virtual void apply_to_buf(char* buf_data);
 
 protected:
     virtual void serialize_data(char* destination) const;
-    virtual size_t get_data_size() const;
+    virtual uint16_t get_data_size() const;
 
 private:
-    size_t src_offset;
-    size_t dest_offset;
-    size_t n;
+    uint16_t src_offset;
+    uint16_t dest_offset;
+    uint16_t n;
 };
 
 #endif	/* __BUF_PATCH_HPP__ */
